# [Nested] Optimize List Type in `list_value`