--- a/test/sql/function/list/list_value.test
+++ b/test/sql/function/list/list_value.test
@@ -12,11 +12,6 @@ SELECT LIST_VALUE([1, 7], [2], [3]);
 [[1, 7], [2], [3]]
 
 query I
-SELECT LIST_VALUE([1, 7], [2], [3]);
-----
-[[1, 7], [2], [3]]
-
-query I
 SELECT LIST_VALUE([1, 7], [2], [3], NULL);
 ----
 [[1, 7], [2], [3], NULL]