--- a/test/sql/function/list/list_value.test_slow
+++ b/test/sql/function/list/list_value.test_slow
@@ -1,0 +1,24 @@
+# name: test/sql/function/list/list_value.test_slow
+# description: Test the list_value function
+# group: [list]
+
+test very large table
+statement ok
+CREATE TABLE large_list_table(a INTEGER[], b INTEGER[], c INTEGER[]);
+
+statement ok
+INSERT INTO large_list_table SELECT [i, i, i], [i + 1, i + 1], [i + 2] FROM range(10000) tbl(i);
+
+query I
+SELECT LIST_VALUE(a, b, c) FROM large_list_table WHERE a[1] = 1;
+----
+[[1, 1, 1], [2, 2], [3]]
+
+load
+CREATE TABLE large_list_table AS SELECT [i, i, i] AS a, [i + 1, i + 1] AS b, [i + 2] AS c FROM range(100000000) tbl(i);
+
+run
+SELECT LIST_VALUE(a, b, c) FROM large_list_table WHERE a[1] = 1;
+
+result I
+[[1, 1, 1], [2, 2], [3]]