In https://github.com/duckdb/duckdb/pull/12468 `list_value` was optimized for primitive types. This PR aims to build on that and optimize list types.

For example:

##### Large Tables
```sql
CREATE TABLE large_list_table AS SELECT [i, i, i] AS a, [i + 1, i + 1] AS b, [i + 2] AS c FROM range(100000000) tbl(i);

SELECT LIST_VALUE(a, b, c) FROM large_list_table;
```

| 1.2.2 | New |
|---|---|
| 28.55s |  8.68s |

##### Large Lists
```sql
CREATE TABLE large_list AS SELECT list(i) AS a FROM range(1000000) t(i);

SELECT list_value(a, a, a, a, a) FROM large_list;
```
| 1.2.2 | New |
|---|---|
| 0.487s |  0.0234s |

##### Nested Lists
```sql
CREATE TABLE nested_lists AS SELECT [[i], [i + 1]] AS a, [[i, i], [i + 1, i + 1]] as b FROM range(10000) t(i);

SELECT list_value(a, b, a, b, a, b, a, b, a, b, a, b, a, b) FROM nested_lists;
```

| 1.2.2 | New |
|---|---|
| 0.128s |  0.0075s |

While these results show improvements, the timings are still slower than desired. Profiling suggests that most of the time is spent in VectorOperations::Copy. Any feedback or suggestions on how to further improve performance would be greatly appreciated!

Some additional tests and benchmarks have also been included.